Move slim RareGame to different file, to avoid circular import
